Discard .rdata$zzz sub-section in genlds output
GNU as generates this section from .ident pseudo-ops that are used
to identify the compiler which generated the object file.

Usually these go into a .comment section in ELF objects but for
COFF, they are lumped to .rdata and cumulate instead of merge.

We don't want these, so stripping out.

Credits to @Ericson2314 for finding a way to remove them from the
output executable.